fix sonar finding with cycles in file imports (part 2)
