+++ packages/app-config-writer/test/unit/preview-config/ui5-yaml.test.ts
@@ -2,11 +2,12 @@ import { create, type Editor } from 'mem-fs-editor';
 import { create as createStorage } from 'mem-fs';
 import { join } from 'path';
 import { ToolsLogger } from '@sap-ux/logger';
-import { updatePreviewMiddlewareConfigs, updatePreviewMiddlewareConfig } from '../../../src/preview-config/ui5-yaml';
+import { updatePreviewMiddlewareConfigs } from '../../../src/preview-config/ui5-yaml';
 import type { PreviewConfigOptions } from '../../../src/types';
 import type { CustomMiddleware } from '@sap-ux/ui5-config';
 import * as projectAccess from '@sap-ux/project-access';
 import type { Script } from '../../../src/common/package-json';
+import { updatePreviewMiddlewareConfig } from '../../../src/common/ui5-yaml';
 
 describe('update preview middleware config', () => {
     const logger = new ToolsLogger();
